Shop sphere

+++ src/main/java/com/Ecommerce/ShopSphere/Service/CartService.java
@@ -10,6 +10,7 @@ import org.springframework.stereotype.Service;
 import com.Ecommerce.ShopSphere.DTO.Cart.AddToCartDto;
 import com.Ecommerce.ShopSphere.DTO.Cart.CartDto;
 import com.Ecommerce.ShopSphere.DTO.Cart.CartItemsResponse;
+import com.Ecommerce.ShopSphere.Exceptions.CustomException;
 import com.Ecommerce.ShopSphere.Exceptions.InvalidCartId;
 import com.Ecommerce.ShopSphere.Model.Cart;
 import com.Ecommerce.ShopSphere.Model.Product;
@@ -48,10 +49,14 @@ public class CartService {
         return cartItemResponse;
     }
 
-    public void deleteCartItem(Integer cartId) {
+    public void deleteCartItem(Integer cartId,User user) {
         Optional<Cart> cartItem = cartRepo.findById(cartId);
         if(cartItem.isEmpty()){
             throw new InvalidCartId("There is no cart item present with id: "+cartId);
+        }
+        Cart item = cartItem.get();
+        if(item.getUser() != user){
+            throw new CustomException("CartItem does not belongs to the user");
         }
         cartRepo.deleteById(cartId);
     }
